+++ rust/compaction/src/datafusion.rs
@@ -26,10 +26,7 @@ use crate::{
 };
 use arrow::util::pretty::pretty_format_batches;
 use datafusion::{
-    datasource::{
-        file_format::{format_as_file_type, parquet::ParquetFormatFactory},
-        physical_plan::ParquetExec,
-    },
+    datasource::file_format::{format_as_file_type, parquet::ParquetFormatFactory},
     error::DataFusionError,
     execution::{
         config::SessionConfig, context::SessionContext, options::ParquetReadOptions,
@@ -37,36 +34,18 @@ use datafusion::{
     },
     logical_expr::{LogicalPlanBuilder, ScalarUDF, SortExpr},
     parquet::basic::{BrotliLevel, GzipLevel, ZstdLevel},
-    physical_plan::{
-        accept, collect,
-        filter::FilterExec,
-        metrics::{MetricValue, MetricsSet},
-        projection::ProjectionExec,
-        ExecutionPlan, ExecutionPlanVisitor,
-    },
+    physical_plan::{accept, collect},
     prelude::*,
 };
 use log::{error, info, warn};
+use metrics::{log_metrics, RowCounts};
 use num_format::{Locale, ToFormattedString};
-use std::{collections::HashMap, io::Write, iter::once, sync::Arc};
+use std::{collections::HashMap, iter::once, sync::Arc};
 use url::Url;
 
+mod metrics;
 pub mod sketch;
 mod udf;
-
-/// Per file metrics to collect from the Parquet reader after compaction
-pub const FILE_METRICS: [&'static str; 10] = [
-    "metadata_load_time",
-    "statistics_eval_time",
-    "page_index_eval_time",
-    "bytes_scanned",
-    "row_groups_pruned_statistics",
-    "row_groups_matched_statistics",
-    "row_groups_pruned_bloom_filter",
-    "row_groups_matched_bloom_filter",
-    "page_index_rows_pruned",
-    "page_index_rows_matched",
-];
 
 /// Starts a Sleeper compaction.
 ///
@@ -252,100 +231,6 @@ async fn collect_stats(
     accept(physical_plan.as_ref(), &mut stats)?;
     log_metrics(&stats.file_metrics);
     Ok(stats)
-}
-
-/// Log some collected per file metrics
-fn log_metrics(metric_map: &HashMap<Url, Vec<MetricValue>>) {
-    for (file, metrics) in metric_map {
-        let mut formatted_metrics = Vec::new();
-        for m in metrics {
-            write!(&mut formatted_metrics, "{} = {}, ", m.name(), m)
-                .expect("Write metrics to string failed: This is a bug");
-        }
-        info!(
-            "File {} {{ {} }}",
-            file.as_str(),
-            String::from_utf8(formatted_metrics).expect("Error in UTF-8 parsing! This is a bug")
-        );
-    }
-}
-
-/// Simple struct used for storing the collected statistics from an execution plan.
-struct RowCounts {
-    pub rows_read: usize,
-    pub rows_written: usize,
-    pub file_metrics: HashMap<Url, Vec<MetricValue>>,
-}
-
-impl RowCounts {
-    /// Pre-populates the map of file metrics from the given list of files
-    pub fn new(inputs: &[Url]) -> Self {
-        Self {
-            rows_read: 0,
-            rows_written: 0,
-            // Populate map with names
-            file_metrics: HashMap::from_iter(inputs.iter().map(|n| (n.to_owned(), vec![]))),
-        }
-    }
-}
-
-impl From<&RowCounts> for CompactionResult {
-    fn from(value: &RowCounts) -> Self {
-        Self {
-            rows_read: value.rows_read,
-            rows_written: value.rows_written,
-        }
-    }
-}
-
-impl ExecutionPlanVisitor for RowCounts {
-    type Error = DataFusionError;
-
-    fn pre_visit(&mut self, plan: &dyn ExecutionPlan) -> Result<bool, Self::Error> {
-        // read output records from final projection stage
-        let maybe_projection = plan
-            .as_any()
-            .downcast_ref::<ProjectionExec>()
-            .and_then(ExecutionPlan::metrics);
-        // read input records from filter stage, not parquet read stage, since parquet
-        // read stage may not filter precisely to range needed and therefore over-reports
-        // number of rows read
-        let maybe_parq_read = plan
-            .as_any()
-            .downcast_ref::<FilterExec>()
-            .and_then(ExecutionPlan::metrics);
-        if let Some(m) = maybe_projection {
-            self.rows_written = m.output_rows().unwrap_or_default();
-        }
-        if let Some(m) = maybe_parq_read {
-            self.rows_read = m.output_rows().unwrap_or_default();
-        }
-        // Read other metrics on a per file basis
-        if let Some(m) = plan
-            .as_any()
-            .downcast_ref::<ParquetExec>()
-            .and_then(ExecutionPlan::metrics)
-        {
-            for (file, file_metrics) in self.file_metrics.iter_mut() {
-                for metric_name in FILE_METRICS {
-                    if let Some(val) = metric_sum_by_filename(&m, metric_name, file.as_str()) {
-                        file_metrics.push(val);
-                    }
-                }
-            }
-        }
-        Ok(true)
-    }
-}
-
-/// Sum all metrics of a given name for a given filename.
-fn metric_sum_by_filename(metrics: &MetricsSet, name: &str, filename: &str) -> Option<MetricValue> {
-    metrics.sum(|m| {
-        m.value().name() == name
-            && m.labels()
-                .iter()
-                .any(|label| label.name() == "filename" && filename.ends_with(label.value()))
-    })
 }
 
 /// Create the `DataFusion` filtering expression from a Sleeper region.
